+++ bin/varnishd/http1/cache_http1_fetch.c
@@ -74,7 +74,7 @@ V1F_SendReq(struct worker *wrk, struct busyobj *bo, uint64_t *ctr_hdrbytes,
 	struct http *hp;
 	int j;
 	ssize_t i;
-	uint64_t bytes, hdrbytes;
+	uint64_t bytes = 0, hdrbytes;
 	struct http_conn *htc;
 	int do_chunked = 0;